README.md

+++ pertemuan-13/read.php
@@ -46,9 +46,9 @@
     <tr>
       <td><?= $i++ ?></td>
       <td>
-      <a href="edit.php?cid=<?= (int)$row['cid']; ?>">Edit</a>
-      <a href="proses_delete.php">Delete</a>
-    </td>
+        <a href="edit.php?cid=<?= (int)$row['cid']; ?>">Edit</a>
+        <a onclick="return confirm('Hapus?')" href="proses_delete.php?cid=<?= (int)$row['cid']; ?>">Delete</a>
+      </td>
       <td><?= $row['cid']; ?></td>
       <td><?= htmlspecialchars($row['cnama']); ?></td>
       <td><?= htmlspecialchars($row['cemail']); ?></td>
